Refactor Numa to LPAR affinity
